% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_fpcr.R
\name{cv_fpcr}
\alias{cv_fpcr}
\title{Crossvalidated Functional Principal Components Regression using 2D Finite Elements basis}
\usage{
cv_fpcr(
  X,
  Y,
  ncomp = min(10, ncol(X)),
  center = TRUE,
  basisobj,
  penalty_vec = c(1e-04, 0.5, 1, 10),
  PCAvalidation = "KFold",
  NFolds = 5,
  folds = 5,
  verbose = TRUE,
  stripped = TRUE,
  ...
)
}
\arguments{
\item{X}{a number-of-observations times nodes matrix.}

\item{Y}{a number-of-observations times reponses matrix.}

\item{ncomp}{number of components, integer.}

\item{center}{center the data.}

\item{basisobj}{a Finite Elements basis as in the fdaPDE package.}

\item{penalty_vec}{a vector of possible penalties.}

\item{PCAvalidation}{validation method in \code{fdaPDE::FPCA.FEM}.}

\item{NFolds}{number of folds to use in \code{fdaPDE::FPCA.FEM}.}

\item{folds}{a user defined list of folds (as generated by caret::createFolds())
or an integer indicating the number of folds. Independent of
\code{fdaPDE::FPCA.FEM}'s validation.}

\item{verbose}{logical, indicating if messages should be printed.}

\item{stripped}{logical.  If \code{TRUE} the calculations are stripped as
much as possible for speed. Particularly, if \code{FALSE} (default) it computes
the final models using the best combination of penalties.
Inspired by package \code{pls}.}

\item{...}{further arguments.  Currently not used}
}
\value{
A list of crossvalidates erros (CVEs) and penalties giving the minimum
CVEs per number of components.
}
\description{
Crossvalidated Functional Principal Components Regression using 2D Finite Elements basis
}
\examples{
# 2D example:

# Generate data (30 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 30, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
FEM_basis <- L[["basisobj"]]

cv_fpcr <- cv_fpcr(X = X, Y = Y, penalty_vec = c(0.01, 10, 100), center = TRUE,
                  ncomp = 3, folds = 5, PCAvalidation = "KFold", NFolds = 5,
                  basisobj = FEM_basis, verbose = TRUE, stripped = FALSE)

final_fpcr <- cv_fpcr$final_model
}
