% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcr_fem.R
\name{fpcr_fem}
\alias{fpcr_fem}
\title{Functional Principal Components Regression using 2D Finite Elements basis}
\usage{
fpcr_fem(
  X,
  Y,
  ncomp = 3,
  center = TRUE,
  basisobj,
  penalty_vec = c(1e-04, 0.5, 1, 10),
  PCAvalidation = "KFold",
  NFolds = 5,
  verbose = TRUE,
  stripped = FALSE,
  ...
)
}
\arguments{
\item{X}{a number of observations times nodes matrix.}

\item{Y}{a number of observations times reponses matrix.}

\item{ncomp}{number of components, integer.}

\item{center}{logical, indicating if data should be centered.}

\item{basisobj}{a Finite Elements basis as in the \code{fdaPDE} package.}

\item{penalty_vec}{a vector of of penalties.}

\item{PCAvalidation}{validation method in \code{fdaPDE::FPCA.FEM}.}

\item{NFolds}{number of folds to use in \code{fdaPDE::FPCA.FEM}.}

\item{verbose}{logical, indicating if messages should be printed.}

\item{stripped}{logical.  If \code{TRUE} the calculations are stripped as
much as possible for speed; this is meant for use with cross-validation or
simulations when only the coefficients are needed.  Defaults to
\code{FALSE}. Inspired by package \code{pls}.}

\item{...}{further arguments.  Currently not used}
}
\value{
an fpcr_fem model.
}
\description{
Functional Principal Components Regression using 2D Finite Elements basis
}
\examples{
# Generate data (50 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 50, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
FEM_basis <- L[["basisobj"]]
pc_res <- fpcr_fem(X, Y, ncomp = 3, center = TRUE, basisobj = FEM_basis,
                   penalty_vec = 10^seq(-2, 4, length.out = 5),
                   PCAvalidation = "KFold", NFolds = 5,
                   stripped = FALSE )
}
