% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fpcr_fem}
\alias{predict.fpcr_fem}
\title{Predict new data for FPCR-FEM model}
\usage{
\method{predict}{fpcr_fem}(object, newdata, ...)
}
\arguments{
\item{object}{an \code{fpcr_fem} model.}

\item{newdata}{new data matrix.}

\item{...}{further arguments.  Currently not used}
}
\value{
predicted values
}
\description{
Predict new data for FPCR-FEM model
}
\examples{
#' # Generate data (50 samples, 100 nodes):
x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

L <- generate_2d_data(x, y, 50, 3, 0.95)

X <- L[["X"]]
Y <- L[["Y"]]
FEM_basis <- L[["basisobj"]]

pc_res <- fpcr_fem(X, Y, ncomp = 3, center = TRUE, basisobj = FEM_basis,
                   penalty_vec = 10^seq(-2, 4, length.out = 5),
                   PCAvalidation = "KFold", NFolds = 5,
                   stripped = FALSE )

Y_pred <- predict(pc_res, X)
Y_hat <- fitted.values(pc_res)
plot(Y_pred, Y_hat)
}
